requirements/base.txt
# Channels
asgiref==3.2.7
channels==2.4.0
channels_redis==2.4.2
Twisted==20.3.0

# Django
Django==2.2.11  # norot

# Postgres binding
psycopg2-binary==2.8.5

django-environ==0.4.5

# Imaging
Pillow==7.1.1

# Models
django-model-utils==4.0.0
django-countries-plus==1.3.1
django-languages-plus==1.1.1

# Unicode slugification
awesome-slugify==1.6.5

# Crypto/Hash
argon2-cffi==19.2.0

# Time zones support
pytz==2019.3

# Redis support
django-redis==4.11.0
redis==3.4.1

# Graphene
graphql-core==2.2.0
graphql-relay==2.0.1  # norot
graphene==2.1.8
-e git+ssh://git@gitlab.com/zetadelta/ery/graphene-django.git#egg=graphene-django
django-filter==2.2.0

# Google storage/datastore
google-api-core==1.17.0
google-cloud==0.34.0
google-cloud-datastore==1.12.0
google-cloud-storage==1.27.0
google-resumable-media[requests]==0.5.0

# Reversion
django-reversion==3.0.7

# Serialization
djangorestframework==3.11.0
djangorestframework-xml==2.0.0
djangorestframework-recursive==0.1.2
drf_writable_nested==0.6.0

# JWT
python-jose==3.1.0
PyJWT==1.7.1
django-graphql-jwt==0.2.1  # norot

# XML
lxml==4.5.0

# gRPC
grpcio==1.28.1

# Google Pub/Sub
google-cloud-pubsub==0.38
protobuf==3.11.3

fastnumbers==3.0.0

urllib3==1.24.3  # norot

# Template
jinja2==2.11.2
# Math
pandas==1.0.3
google-auth-oauthlib==0.3.0  # norot
pygsheets==2.0.3.1

# XXX: should be moved to web_runner but kept here to simplify migrations
social-auth-app-django==3.1.0
